Fix node list disappear when running the software from command line.

+++ scripts/file_functions/file_functions.gml
@@ -243,6 +243,7 @@ function filename_combine(_p1, _p2) {
 	////- Actions
 
 function file_copy_override(src, dest) {
+	if(!file_exists(src)) return;
 	if(file_exists_empty(dest)) file_delete(dest);
 	file_copy(src, dest);
 }
